232: move "interpolate" code to separate file
